Add limitation for unauthorized users

+++ src/main/java/com/example/web/controllers/BooksController.java
@@ -12,19 +12,24 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.List;
 import java.util.ArrayList;
+import javax.servlet.http.HttpSession;
 
 @Controller
 public class BooksController {
-    @Autowired
-    private BookRepo bookRepo;
+    private final BookRepo bookRepo;
+
+    public BooksController(BookRepo bookRepo) {
+        this.bookRepo = bookRepo;
+    }
 
     @GetMapping("/")
     public String start() {
-        return "redirect:/shop?page=1";
+        return "redirect:/signin";
     }
 
     @GetMapping("/shop")
-    public String mainPage(@RequestParam Integer page, Map<String, Object> model) {
+    public String mainPage(@RequestParam Integer page,
+                           HttpSession session, Map<String, Object> model) {
         Pageable paginator = PageRequest.of(page-1, 5);
 
         Iterable<Book> books = bookRepo.findAll(paginator);
@@ -48,33 +53,52 @@ public class BooksController {
         model.put("pages", pages);
         model.put("number", page);
         model.put("count", count);
+        model.put("session", session);
 
         return "main";
     }
 
     @GetMapping("/create")
-    public String addBookForm() {
+    public String addBookForm(Map<String, Object> model, HttpSession session) {
+        if (session.getAttribute("role") == null)
+            return "redirect:/signin";
+
+        Book book = new Book();
+        book.setName("");
+        book.setAuthor("");
+        book.setPrice(0);
+        model.put("book", book);
+        model.put("session", session);
+
         return "create";
     }
 
     @PostMapping("/create")
     public String addBook(@RequestParam String name,
-                      @RequestParam String author,
-                      @RequestParam Integer price) {
+                          @RequestParam String author,
+                          @RequestParam Integer price,
+                          Map<String, Object> model, HttpSession session) {
         Book book = new Book(name, author, price);
         bookRepo.save(book);
+        model.put("session", session);
 
         return "redirect:/shop?page=1";
     }
 
     @GetMapping("/edit")
-    public String editBookForm(@RequestParam Integer id, Map<String, Object> model) {
-        Optional<Book> book = bookRepo.findById(id);
+    public String editBookForm(@RequestParam Integer id,
+                               Map<String, Object> model, HttpSession session) {
+        if (session.getAttribute("role") == "superuser") {
+            Optional<Book> book = bookRepo.findById(id);
 
-        if (book.isPresent()) {
-            model.put("book", book.get());
-            return "edit";
+            if (book.isPresent()) {
+                model.put("book", book.get());
+                model.put("session", session);
+
+                return "edit";
+            }
         }
+
         return "redirect:/shop?page=1";
     }
 
@@ -82,20 +106,23 @@ public class BooksController {
     public String editBook(@RequestParam Integer id,
                            @RequestParam String name,
                            @RequestParam String author,
-                           @RequestParam Integer price) {
+                           @RequestParam Integer price,
+                           Map<String, Object> model, HttpSession session) {
         Book book = bookRepo.findById(id).get();
         book.setName(name);
         book.setAuthor(author);
         book.setPrice(price);
         bookRepo.save(book);
+        model.put("session", session);
 
         return "redirect:/shop?page=1";
     }
 
     @GetMapping("/delete")
-    public String deleteBook(@RequestParam Integer id) {
-        if (bookRepo.findById(id).isPresent())
-            bookRepo.deleteById(id);
+    public String deleteBook(@RequestParam Integer id, HttpSession session) {
+        if (session.getAttribute("role") == "superuser")
+            if (bookRepo.findById(id).isPresent())
+                bookRepo.deleteById(id);
 
         return "redirect:/shop?page=1";
     }
